XWIKI-19732: Replace ActiveInstalls by ActiveInstalls2 in XS * Add a docker-based functional test to prove that a ping is sent when XS starts * Also upgrade LO to 7.2.7.2 (LTS) * Also upgrade TC to 1.17.2 * Add the ability for functional docker-based tests to inject dynamically-generated TestConfiguration (needed to inject the activeinstall ping URL)

+++ xwiki-platform-core/xwiki-platform-activeinstalls2/xwiki-platform-activeinstalls2-api/src/main/java/org/xwiki/activeinstalls2/internal/data/DocumentsPing.java
@@ -29,16 +29,16 @@ import java.util.List;
  */
 public class DocumentsPing
 {
-    private List<Integer> all;
+    private List<Long> all;
 
-    private int main;
+    private long main;
 
-    private int total;
+    private long total;
 
     /**
      * @return the number of documents in the main wiki
      */
-    public int getMain()
+    public long getMain()
     {
         return this.main;
     }
@@ -46,7 +46,7 @@ public class DocumentsPing
     /**
      * @param documentCount see {@link #getMain()}
      */
-    public void setMain(int documentCount)
+    public void setMain(long documentCount)
     {
         this.main = documentCount;
     }
@@ -54,7 +54,7 @@ public class DocumentsPing
     /**
      * @return the total number of documents in the main wiki and all the sub wikis
      */
-    public int getTotal()
+    public long getTotal()
     {
         return this.total;
     }
@@ -62,7 +62,7 @@ public class DocumentsPing
     /**
      * @param documentCount see {@link #getTotal()}
      */
-    public void setTotal(int documentCount)
+    public void setTotal(long documentCount)
     {
         this.total = documentCount;
     }
@@ -71,7 +71,7 @@ public class DocumentsPing
      * @return the number of documents in each subwiki, excluding the main wiki, as an array
      *         (we don't want to expose wiki names since that would contain private information)
      */
-    public List<Integer> getWikis()
+    public List<Long> getWikis()
     {
         return this.all;
     }
@@ -79,7 +79,7 @@ public class DocumentsPing
     /**
      * @param documentCount see {@link #getWikis()}
      */
-    public void setWikis(List<Integer> documentCount)
+    public void setWikis(List<Long> documentCount)
     {
         this.all = documentCount;
     }
